CDAP-12617 Split SparkTestRun into multiple test classes.
